moved InvalidSessionException to exception package
