+++ components/org.wso2.carbon.identity.sso.agent/src/main/java/org/wso2/carbon/identity/sso/agent/exception/InvalidSessionException.java
@@ -16,9 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.sso.agent;
-
-import org.wso2.carbon.identity.sso.agent.exception.SSOAgentException;
+package org.wso2.carbon.identity.sso.agent.exception;
 
 /**
  * Invalid session exception.
